requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
ipython==9.9.0
alembic==1.18.0
passlib==1.7.4
argon2_cffi==25.1.0
redis==5.2.1
livekit-api==1.1.0